Deleted old typedefs, resolvers, queries and mutations after testing new ones. renamed version 2 of each file and updated import file paths.
