+++ client/src/components/SignupBox.jsx
@@ -1,7 +1,7 @@
 import { Alert } from "react-bootstrap";
 import { useState } from "react";
 import { useMutation } from "@apollo/client";
-import { ADD_USER } from "../utils/mutations2";
+import { ADD_USER } from "../utils/mutations";
 import Auth from "../utils/auth";
 /* eslint-disable no-unused-vars */
