+++ client/src/utils/queries.js
@@ -1,22 +1,26 @@
-// import { gql } from "@apollo/client";
+import { gql } from "@apollo/client";
 
-// export const QUERY_ME = gql`
-//   {
-//     me {
-//       _id
-//       username
-//       email
-//       password
-//       savedMovieCount
-//       savedMovies {
-//         movieId
-//         title
-//         image
-//         directors
-//         review
-//         rating
-//         favorite
-//       }
-//     }
-//   }
-// `;
+export const QUERY_ME = gql`
+  query Me {
+    me {
+      _id
+      username
+      email
+      password
+      ratedMovieCount
+      ratedMovies {
+        movieId
+        title
+        image
+        rating
+      }
+      favoriteMovieCount
+      favoriteMovies {
+        movieId
+        title
+        image
+        rating
+      }
+    }
+  }
+`;
