+++ server/schemas/typedefs.js
@@ -1,36 +1,46 @@
-// const typeDefs = `
-//     type User {
-//         _id: ID!
-//         username: String!
-//         email: String!
-//         password: String!
-//         savedMovieCount: Int
-//         savedMovies: [Movie]
-//     }
+const typeDefs = `
+    type User {
+        _id: ID!
+        username: String!
+        email: String!
+        password: String!
+        ratedMovieCount: Int
+        ratedMovies: [Movie]
+        favoriteMovieCount: Int
+        favoriteMovies: [Movie]
+    }
 
-//     type Movie {
-//         movieId: String!
-//         title: String!
-//         image: String!
-//         rating: Float!
-//         favorite: Boolean
-//     }
+    type Movie {
+        movieId: String!
+        title: String!
+        image: String!
+        rating: Float!
+    }
 
-//     type Auth {
-//         token: ID!
-//         user: User
-//     }
+    input MovieInput {
+        movieId: String!
+        title: String!
+        image: String!
+        rating: Float!
+    }
 
-//     type Query {
-//         me: User
-//     }
+    type Auth {
+        token: ID!
+        user: User
+    }
 
-//     type Mutation {
-//         addUser(username: String!, email: String!, password: String!): Auth
-//         login(email: String!, password: String!): Auth
-//         addMovie(movieId: String!, title: String!, image: String!, rating: Float!, favorite: Boolean): User
-//         removeMovie(movieId: String!): User
-//     }
-// `;
+    type Query {
+        me: User
+    }
 
-// module.exports = typeDefs;
+    type Mutation {
+        addUser(username: String!, email: String!, password: String!): Auth
+        login(email: String!, password: String!): Auth
+        rateMovie(movieInput: MovieInput): User
+        unRateMovie(movieId: String!): User
+        addMovieToFavorite(movieInput: MovieInput): User
+        unFavoriteMovie(movieId: String!): User
+    }
+`;
+
+module.exports = typeDefs;
